Keep context processor for freebasics instead of templatetag
I've been thinking about this and I think a context processor is
better than a template tag for this. With the tag you have to import
and assign the result inside the template, and we use it quite frequently.

With the context processor it's always available and the overhead of checking
the request headers doesn't feel huge to me. I think it's probably quite quick.
