+++ gem/templates/springster/core/tags/bannerpages.html
@@ -1,6 +1,5 @@
 {% load gem_tags %}
  {% if bannerpages %}
- {% is_via_freebasics as is_via_freebasics %}
    {% for bannerpage in bannerpages %}
    {% if bannerpage.hide_on_freebasics %}
      {% if not is_via_freebasics %}
